feature(common): Changing the name of the components from Spanish to English
