+++ src/App.js
@@ -1,16 +1,16 @@
 import Theme from "./Theme/Theme";
-import HistorialPedidos from "./ui/components/HistorialPedidos/HistorialPedidos";
-import Preferences from "./ui/components/Preferences/Preferences";
+import ordersHistory from "./ui/components/ordersHistory/ordersHistory";
+import Preferences from "./ui/components/orderPreferences/orderPreferences";
 import NavMenu from "./ui/components/navMenu/NavMenu";
-import "bootstrap/dist/css/bootstrap.min.css";
+// import "bootstrap/dist/css/bootstrap.min.css";
 import Cliente from "./Cliente"
 
 function App() {
   return (
     <Theme>
-      <HistorialPedidos/>
-    <NavMenu />
-   <Cliente/>
+      <ordersHistory/>
+      <NavMenu />
+      <Cliente/>
     </Theme>
   );
 }
